Correct linter errors
