--- Teacher.rb
@@ -1,18 +1,0 @@
-require "./Person.rb"
-
-class Student < Person
-    #Define constructor
-    def initialize(specialization)
-        #Call attributes from parent
-        super(age, name = "Unknown", parent_permission = true)
-        @specialization = specialization
-    end
-
-    #Define getter and setter for @specialization
-    attr_accessor :specialization
-
-    #Override can_use_services? method
-    def can_use_services?
-        return true
-    end
-end
